Refactor: service.py 분리#22
